// src/app/shared/infrastructure/api/generated/model/driverReputationResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Backend-WebApplication
 * ChapaTuRuta REST API â Identity & Access Management, Driver Management, and Monetization
 * OpenAPI spec version: 1.0.0
 */

export interface DriverReputationResponse {
  driverId?: string;
  averageScore?: number;
  totalRatings?: number;
}